requirement.txt
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
nltk
pandas
